+++ src/api/method/get_indexer_health.rs
@@ -6,7 +6,7 @@ use super::super::error::PhotonApiError;
 use super::utils::Context;
 
 // TODO: Make this an environment variable.
-const HEALTH_CHECK_SLOT_DISTANCE: i64 = 5;
+pub const HEALTH_CHECK_SLOT_DISTANCE: i64 = 5;
 
 // TODO: Make sure that get_indexer_health formatting matches the Solana RPC formatting.
 pub async fn get_indexer_health(
