+++ src/ingester/fetchers/mod.rs
@@ -1,2 +1,3 @@
-pub mod poller;
 pub mod grpc;
+pub mod poller;
+pub mod pollerv2;
